+++ docs/src/examples/components/Menu/Types/MenuExamplePrimary.shorthand.tsx
@@ -7,6 +7,6 @@ const items = [
   { key: 'events', content: 'Upcoming Events' },
 ]
 
-const MenuExamplePrimary = () => <Menu defaultActiveIndex={0} items={items} type="primary" />
+const MenuExamplePrimary = () => <Menu defaultActiveIndex={0} items={items} primary />
 
 export default MenuExamplePrimary